Might as well make the loop indices just guint; the generated code's unlikely to be worse, and could be better on the 32-bit-and-bigger machines on which Wireshark runs.
svn path=/trunk/; revision=42116

+++ epan/dissectors/packet-bthci_evt.c
@@ -1270,7 +1270,7 @@ static int
 dissect_bthci_evt_ext_inquiry_response(tvbuff_t *tvb, int offset, packet_info *pinfo _U_, proto_tree *tree)
 {
     guint8      length, type;
-    guint16     i, j;
+    guint       i, j;
     proto_item *ti_eir         = NULL;
     proto_item *ti_eir_subtree = NULL;
 
@@ -1301,7 +1301,7 @@ dissect_bthci_evt_ext_inquiry_response(tvbuff_t *tvb, int offset, packet_info *p
                 case 0x02: /* 16-bit Service Class UUIDs, incomplete list */
                 case 0x03: /* 16-bit Service Class UUIDs, complete list */
                     j=0;
-                    while(j<(length-1))
+                    while(j<(guint)(length-1))
                     {
                         proto_tree_add_item(ti_eir_struct_subtree, hf_bthci_evt_sc_uuid16, tvb, offset+i+j+2, 2, ENC_LITTLE_ENDIAN);
                         j+=2;
@@ -1310,7 +1310,7 @@ dissect_bthci_evt_ext_inquiry_response(tvbuff_t *tvb, int offset, packet_info *p
                 case 0x04: /* 32-bit Service Class UUIDs, incomplete list */
                 case 0x05: /* 32-bit Service Class UUIDs, complete list */
                     j=0;
-                    while(j<(length-1))
+                    while(j<(guint)(length-1))
                     {
                         proto_tree_add_item(ti_eir_struct_subtree, hf_bthci_evt_sc_uuid32, tvb, offset+i+j+2, 4, ENC_LITTLE_ENDIAN);
                         j+=4;
@@ -1319,7 +1319,7 @@ dissect_bthci_evt_ext_inquiry_response(tvbuff_t *tvb, int offset, packet_info *p
                 case 0x06: /* 128-bit Service Class UUIDs, incomplete list */
                 case 0x07: /* 128-bit Service Class UUIDs, complete list */
                     j=0;
-                    while(j<(length-1))
+                    while(j<(guint)(length-1))
                     {
                         proto_tree_add_item(ti_eir_struct_subtree, hf_bthci_evt_sc_uuid128, tvb, offset+i+j+2, 16, ENC_NA);
                         j+=16;
